+++ StatsData.h
@@ -20,9 +20,9 @@
 #include "Config.h"
 
 typedef struct {
-	int16_t avg;
-	int16_t min;
-	int16_t max;
+	int8_t avg;
+	int8_t min;
+	int8_t max;
 	uint8_t day; // day number in history. 0 - now, 1 - yesterday, 2 - before yesterday, and so on.
 } Temp;
